--- SemiProject/src/main/java/com/kh/member/model/MemberService/MemberService.java
@@ -1,11 +1,0 @@
-package com.kh.member.model.MemberService;
-
-import com.kh.member.model.vo.Member;
-
-public interface MemberService {
-	
-	Member loginMember(Member m);
-	int insertMember(Member m);
-	Member searchMember(Member m);
-	Member searchMemberPwd(Member m);
-}
