+++ SemiProject/src/main/java/com/kh/member/model/dao/MemberDao.java
@@ -10,4 +10,16 @@ public class MemberDao {
 		return sqlSession.selectOne("memberMapper.loginMember", m);
 	}
 	
+	public int insertMember(SqlSession sqlSession, Member m) {
+		int result = sqlSession.insert("memberMapper.insertMember", m);
+		return result;
+	}
+		
+	public Member searchMember(SqlSession sqlSession, Member m) {
+		return sqlSession.selectOne("memberMapper.searchMember", m);
+	}
+	
+	public Member searchMemberPwd(SqlSession sqlSession, Member m) {
+		return sqlSession.selectOne("memberMapper.searchMemberPwd", m);
+	}
 }
