+++ SemiProject/src/main/java/com/kh/member/model/service/MemberService.java
@@ -3,7 +3,9 @@ package com.kh.member.model.service;
 import com.kh.member.model.vo.Member;
 
 public interface MemberService {
-
-	Member loginMember(Member m);
 	
+	Member loginMember(Member m);
+	int insertMember(Member m);
+	Member searchMember(Member m);
+	Member searchMemberPwd(Member m);
 }
